fix split

+++ ft_memset.c
@@ -2,8 +2,8 @@
 
 void *ft_memset(void *dest, int x, size_t len)
 {
-  while (len > -1)
-    ((unsigned char*)dest)[len--] = (unsigned char)x;
+  while ((int)len > -1)
+    ((unsigned char*)dest)[(int)len--] = (unsigned char)x;
   
   return (dest);
 }
